# doc/api.yaml
openapi: 3.0.0
info:
  title: WASAPhoto
  description: |-
    Rimani in contatto con i tuoi amici condividendo le foto dei momenti speciali, grazie a WASAPhoto! 
    Potete caricare le vostre foto direttamente dal vostro PC, che saranno visibili a tutti coloro che vi seguono.
  version: "1"
...
paths:
...
  /users/{username}/ban/{banId}:
    parameters:
    - $ref: "#/components/parameters/username"
    - $ref: "#/components/parameters/banId"

    # endpoint per bannare un utente
    put:
      security:
        - bearerAuth: []
      tags : ["Ban"]

      summary: Bans a user
      description: |-
        Bans a user.
      operationId: banUser
      responses:
        "201": 
          description: User banned successfully.
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/ban"
        "400": { $ref: "#/components/responses/BadRequestError" }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    
    # endpoint per sbannare un utente
    delete:
      security:
        - bearerAuth: []
      tags : ["Ban"]
      summary: Unbans a user
      description: |-
        Unbans a user.
      operationId: unbanUser
      responses:
       "204": { description: Ban deleted successfully. }
       "400": { $ref: "#/components/responses/BadRequestError" }
       "401": { $ref: "#/components/responses/UnauthorizedError" }
       "500": { $ref: "#/components/responses/InternalServerError" }
  /users/{username}/ban:
    # endpoint per ottenere la lista degli utenti bannati
    parameters:
    - $ref: "#/components/parameters/username"
    get:
      security:
        - bearerAuth: []
      tags : ["Ban"]
      summary: returns user's bans
      description: |-
        Returns user's bans
      operationId: getBansList
      responses:
        "200": 
          description: User bans retrived successfully.
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/ban"
        "400": { $ref: "#/components/responses/BadRequestError" }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
        "500": { $ref: "#/components/responses/InternalServerError" }

################################################################
#! FOLLOWERS
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses: 
    UnauthorizedError:
      description: Access token is missing or invalid
    BadRequestError:
      description: The request was not compliant with the documentation (eg. missing fields, etc).
    InternalServerError:
      description: The server encountered an internal error. Further info in server logs.
  parameters:
    username:
      schema:
          description: username.
          type: string
          pattern: '^.*?$'
          example: MikeRossi
          minLength: 3
          maxLength: 30
      name: username
      in: path
      required: true
      description: "username of user."
...
    banId:
      schema:
        description: banId schema.
        type: string
        pattern: '^.*?$'
        example: "#B1234"
        minLength: 3
        maxLength: 16
      name: banId
      in: path
      required: true
      description: "*banId of a single ban on a user."


#! SCHEMAS
  schemas:
...
    ban:
      title: ban
      description: Ban schema.
      type: object

      # le proprietà del ban
      properties:
        bannedId:
          description: user banned identifier.
          type: boolean
          example: true
        
        userId:
          description: Unique user identifier.
          type: string
          pattern: '^.*?$'
          example: "$U1234"
          minLength: 3
          maxLength: 16
        
        banId:
          description: Unique ban identifier.
          type: string
          pattern: '^.*?$'
          example: "$B1234"
          minLength: 3
          maxLength: 16